pyproject.toml
[project]
name = "simple-ai-agent"
version = "0.1.0"
description = "Baremetal AI agent example with tool calls."
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = []
